Hamming code used to implement the ecc algorithm
A hamming code is used to ecc the 32bit data. There are 6 bits parity code and 1 bit odd even parity code. Therefore, the module is able to detect two bits mistakes and correct 1 bit mistake. The only drawback is the violation of timing analysis, which results from only one-stage pipeline. The next step is to optimized the module by two directions: 1. add pipeline stage; 2. simplify the logic of MUX and logic operations.